fix: exclude system-upgrade kustomizations from cluster-config-secret substitution
The system-upgrade-controller-plans kustomization was failing because the parent
cluster-apps kustomization was applying a patch that added cluster-config-secret
as a substitute source, but this secret doesn't contain the KUBERNETES_VERSION
and TALOS_VERSION variables needed for the upgrade plans.

Updated the label selector to specifically exclude the system-upgrade related
kustomizations from getting the cluster-config-secret substitution patch.